+++ py/ai/h2o/sparkling/ml/models/__init__.py
@@ -18,3 +18,5 @@
 from ai.h2o.sparkling.ml.models.H2OTargetEncoderModel import H2OTargetEncoderModel
 from ai.h2o.sparkling.ml.models.H2OTargetEncoderMOJOModel import H2OTargetEncoderMOJOModel
 from ai.h2o.sparkling.ml.models.H2OMOJOSettings import H2OMOJOSettings
+from ai.h2o.sparkling.ml.models.H2OMOJOModel import H2OMOJOModel
+from ai.h2o.sparkling.ml.models.H2OMOJOPipelineModel import H2OMOJOPipelineModel
