+++ scoring/build.gradle
@@ -13,6 +13,7 @@ sourceSets {
 dependencies {
     compileOnly "org.apache.spark:spark-core_${scalaBaseVersion}:${sparkVersion}"
     compileOnly "org.apache.spark:spark-mllib_${scalaBaseVersion}:${sparkVersion}"
+    compileOnly project(":sparkling-water-macros")
 
     compile "ai.h2o:mojo2-runtime-api:${mojoPipelineVersion}"
     compile("ai.h2o:h2o-genmodel:${h2oVersion}")
