+++ scoring/src/main/scala/ai/h2o/sparkling/ml/models/H2OMOJOLoader.scala
@@ -20,7 +20,6 @@ package ai.h2o.sparkling.ml.models
 import java.io.InputStream
 
 import org.apache.hadoop.fs.Path
-import org.apache.spark.ml.h2o.models.H2OMOJOSettings
 import org.apache.spark.ml.util.Identifiable
 import org.apache.spark.sql.SparkSession
