+++ scoring/src/main/scala/ai/h2o/sparkling/ml/params/H2OMOJOAlgoSharedParams.scala
@@ -16,8 +16,8 @@
 */
 package ai.h2o.sparkling.ml.params
 
+import ai.h2o.sparkling.ml.models.H2OMOJOSettings
 import org.apache.spark.expose.Logging
-import org.apache.spark.ml.h2o.models.H2OMOJOSettings
 import org.apache.spark.ml.param._
 
 /**
